refactor(supi-review): merge small modules into parent files
Consolidate three single-purpose modules into their primary consumers:

- format-helpers.ts (formatLevel, formatLocation) inlined into format-content.ts

- runner-types.ts (ReviewProgress, BriefSynthesisRunResult, etc.) merged into types.ts

- review-instruction-blocks.ts (block catalog + accessors) merged into target/packet.ts

All import paths updated, source files deleted. No behavior changes — 104/104 tests pass.
